requirements.txt
groq==0.4.1
fpdf2==2.7.4
python-dotenv==1.0.0
flask==2.3.3